+++ packages/elements/select/examples/02-radio-select.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import { RadioSelect } from './common/components';
+import { RadioSelect } from '../src';
 import { cities } from './common/data';
 
 // data imported for brevity; equal to the options from Single Select example