+++ packages/fabric/editor-confluence-transformer/tests/browser/cxhtml.ts
@@ -51,6 +51,7 @@ import {
   taskList,
   taskItem,
   date,
+  placeholder,
 } from '@atlaskit/editor-test-helpers';
 
 import {
@@ -827,13 +828,13 @@ describe('ConfluenceTransformer: encode - parse:', () => {
     describe('jira issue', () => {
       check(
         'basic',
-        '<p><ac:structured-macro ac:name="jira" ac:schema-version="1" ac:macro-id="a1a887df-a2dd-492b-8b5c-415d8eab22cf"><ac:parameter ac:name="server">JIRA (product-fabric.atlassian.net)</ac:parameter><ac:parameter ac:name="serverId">70d83bc8-0aff-3fa5-8121-5ae90121f5fc</ac:parameter><ac:parameter ac:name="key">ED-1068</ac:parameter></ac:structured-macro></p>',
+        '<p><ac:structured-macro ac:name="jira" ac:schema-version="3" ac:macro-id="a1a887df-a2dd-492b-8b5c-415d8eab22cf"><ac:parameter ac:name="server">JIRA (product-fabric.atlassian.net)</ac:parameter><ac:parameter ac:name="serverId">70d83bc8-0aff-3fa5-8121-5ae90121f5fc</ac:parameter><ac:parameter ac:name="key">ED-1068</ac:parameter></ac:structured-macro></p>',
         doc(
           p(
             confluenceJiraIssue({
               issueKey: 'ED-1068',
               macroId: 'a1a887df-a2dd-492b-8b5c-415d8eab22cf',
-              schemaVersion: '1',
+              schemaVersion: '3',
               server: 'JIRA (product-fabric.atlassian.net)',
               serverId: '70d83bc8-0aff-3fa5-8121-5ae90121f5fc',
             })(),
@@ -856,6 +857,7 @@ describe('ConfluenceTransformer: encode - parse:', () => {
           type: 'image',
         },
       ],
+      schemaVersion: { value: '2' },
     };
     const attrs = {
       extensionType: 'com.atlassian.confluence.macro.core',
@@ -872,37 +874,12 @@ describe('ConfluenceTransformer: encode - parse:', () => {
         }</ac:parameter>`,
     );
 
-    describe('extensions with resource identifiers', () => {
-      const attrs = {
-        extensionType: 'com.atlassian.confluence.macro.core',
-        extensionKey: 'fake',
-        parameters: {
-          macroParams: {
-            src: { value: 'www.google.com' },
-            spaces: { value: 'abc' },
-          },
-          macroMetadata,
-        },
-      };
-      check(
-        'basic',
-        `<ac:structured-macro ac:name="${
-          attrs.extensionKey
-        }" ac:schema-version="1" ac:macro-id="${
-          macroMetadata.macroId.value
-        }"><ac:parameter ac:name="src"><ri:url ri:value="www.google.com" /></ac:parameter><ac:parameter ac:name="spaces"><ri:space ri:space-key="abc" /></ac:parameter><fab:placeholder-url>${
-          macroMetadata.placeholder[0].data.url
-        }</fab:placeholder-url><fab:display-type>BLOCK</fab:display-type></ac:structured-macro>`,
-        doc(extension(attrs)()),
-      );
-    });
-
     describe('inlineExtension', () => {
       check(
         'basic',
         `<ac:structured-macro ac:name="${
           attrs.extensionKey
-        }" ac:schema-version="1" ac:macro-id="${
+        }" ac:schema-version="2" ac:macro-id="${
           macroMetadata.macroId.value
         }">${paramsAsCXHTML}<fab:placeholder-url>${
           macroMetadata.placeholder[0].data.url
@@ -916,7 +893,7 @@ describe('ConfluenceTransformer: encode - parse:', () => {
         'basic',
         `<ac:structured-macro ac:name="${
           attrs.extensionKey
-        }" ac:schema-version="1" ac:macro-id="${
+        }" ac:schema-version="2" ac:macro-id="${
           macroMetadata.macroId.value
         }">${paramsAsCXHTML}<fab:placeholder-url>${
           macroMetadata.placeholder[0].data.url
@@ -930,7 +907,7 @@ describe('ConfluenceTransformer: encode - parse:', () => {
         'basic',
         `<ac:structured-macro ac:name="${
           attrs.extensionKey
-        }" ac:schema-version="1" ac:macro-id="${
+        }" ac:schema-version="2" ac:macro-id="${
           macroMetadata.macroId.value
         }">${paramsAsCXHTML}<fab:placeholder-url>${
           macroMetadata.placeholder[0].data.url
@@ -949,6 +926,15 @@ describe('ConfluenceTransformer: encode - parse:', () => {
       'date node',
       `<time datetime="${iso}"></time>`,
       doc(p(date({ timestamp: parseDate(iso).valueOf() }))),
+    );
+  });
+
+  describe('placeholder', () => {
+    const text = 'Write something...';
+    check(
+      'placeholder node',
+      `<ac:placeholder>${text}</ac:placeholder>`,
+      doc(p(placeholder({ text }))),
     );
   });