+++ packages/elements/avatar/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/avatar
 
+## 9.0.4
+- [patch] updated the repository url to https://bitbucket.org/atlassian/atlaskit-mk-2 [1e57e5a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e57e5a)
+
+## 9.0.3
+
 ## 9.0.2
 - [patch] Remove babel-plugin-react-flow-props-to-prop-types [06c1f08](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06c1f08)