+++ packages/elements/navigation/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaskit/navigation
 
+## 29.0.2
+- [patch] updated the repository url to https://bitbucket.org/atlassian/atlaskit-mk-2 [1e57e5a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e57e5a)
+
+## 29.0.1
+- [patch] fixes AK-4178 , added fix for double color icon in navigation [c6121d6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c6121d6)
+
+## 29.0.0
+- [major] update navigation and item's usage of react-beautiful-dnd from 2.x to 4.x. [aeefa01](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/aeefa01)
+
 ## 28.0.4
 - [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)